refactor(Utils): turned float into double

+++ src/Utils.cpp
@@ -2,7 +2,7 @@
 #include <numbers>
 #include <cmath>
 
-float normalizeAngle(float angle) {
+double normalizeAngle(double angle) {
     angle = remainder(angle, 2 * std::numbers::pi);
     if (angle < 0) {
         angle = 2 * std::numbers::pi + angle;
@@ -10,6 +10,6 @@ float normalizeAngle(float angle) {
     return angle;
 }
 
-float distanceBetweenPoints(float x1, float y1, float x2, float y2) {
+double distanceBetweenPoints(double x1, double y1, double x2, double y2) {
     return sqrt((x2 - x1) * (x2 - x1) + (y2 - y1) * (y2 - y1));
 }
