+++ src/Utils.h
@@ -1,5 +1,5 @@
 #pragma once
 
-float normalizeAngle(float angle);
+double normalizeAngle(double angle);
 
-float distanceBetweenPoints(float x1, float y1, float x2, float y2);
+double distanceBetweenPoints(double x1, double y1, double x2, double y2);
